CM-910: Update Kong for Elasticsearch
